more general CMT analysis functions
